Move helper functions to UC_Utils
